refactor: rename to `woken_pid`
