+++ kernel/src/process/collections/mod.rs
@@ -1,4 +1,4 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-mod pid;
 mod process;
+mod woken_pid;
